setup.py: use a separate build directory for free-threading python
Free-threading Python binaries aren't ABI compatible with the
traditional builds that still have a GIL. But currently the build_temp
directory, which is drgn's incremental build directory, is shared
between both kinds of interpreters. This can result in runtime issues
like segmentation faults if build artifacts are reused. To avoid this,
add a "t" suffix for free-threading build directories. We should only do
this if the "build-temp" option is not set by the user.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>