setup.py: explicitly use default kernel flavor for test -K
Stephen noticed in #415 that if a new stable release is tagged after the
default vmtest build job runs but before another flavor does, we can end
up in a situation where the newest kernel for a given version isn't the
default flavor. This causes `setup.py test -K` to test a non-default
flavor. `-K` is specifically supposed to test the default, so specify it
explicitly.

Reported-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Signed-off-by: Omar Sandoval <osandov@osandov.com>