mm, slab: Fix test failures on kernels with SLOB
In the Fixes commit, as part of the refactor, the
_identify_kernel_address() function started open-coding the
get_slab_object_info() function. Unfortunately, this resulted in the
function failing when CONFIG_SLOB was enabled, since the AttributeError
for accessing "slab.slab_cache" was not caught.

Open-coding was done to avoid the duplicate bounds check for
in_direct_map. In fact, in_direct_map seems like a useful helper itself,
and by factoring it out of _find_containing_slab() and into its callers,
we can simplify things a bit. Since _find_containing_slab() can handle
SLOB, this fixes the test failure too.

Fixes: 48e0c51b ("helper.common.memory: recognize vmap addresses in identify_address()")

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>