drgn.internal.sudohelper: pass privileged command as string instead of path
At Meta, our Python executables are packaged in a way that doesn't leave
the original .py or .pyc files easily accessible or importable. This
makes the sudohelper fail with:

  can't open file '/proc/self/fd/3/drgn/internal/sudohelper.pyc': [Errno 2] No such file or directory

We can work around this by passing the helper command as a string
literal. It's ugly, but it works with our Meta package and with
traditional installations.

Signed-off-by: Omar Sandoval <osandov@osandov.com>