libdrgn: remove drgn_type definition from public header
struct drgn_type is currently defined in drgn.h, and the type getters
are all static inline functions that access its private fields. This
exposes a ton of implementation details. I'm about to do a big overhaul
of struct drgn_type, which made me conscious of the fact that if we ever
get to the point where we promise ABI stability for libdrgn, this would
be a huge blocker for future changes.

We can fix this by moving struct drgn_type to drgn_internal.h and making
the getters regular extern functions. However, I observed a performance
regression from converting the trivial getters to real function calls.
Instead, we can make the getters inline within libdrgn and extern
publicly with some extra work.

Signed-off-by: Omar Sandoval <osandov@osandov.com>