libdrgn: never deduplicate struct, union, class, enum, or function types
We currently deduplicate types that have empty member, enumerator,
parameter, and template parameter lists. In an upcoming change, we are
going to add extra context about where struct, union, class, enum, and
function types came from (namely, the DWARF DIE address). This doesn't
work if we deduplicate those, so let's not anymore. On the bright side,
we get to simplify the drgn_dedupe_type_set hash and equality functions.

Signed-off-by: Omar Sandoval <osandov@osandov.com>