libdrgn: store DWARF DIE that type came from
There are things that we may want to know about a type but that aren't
worth parsing and storing at the time that we create the struct
drgn_type. A few examples are:

- Alignment
- Fully qualified name
- Base classes

An easy and efficient way to support these is to store the address of
the DWARF DIE that a struct drgn_type was parsed from, which can then be
used to parse additional information lazily.

It'd be more general to store an opaque context and a callback table,
but there's no point in paying the additional overhead for that until we
have a use case for it. For now, we'll just add this layer violation.

Signed-off-by: Omar Sandoval <osandov@osandov.com>