libdrgn: platform: add alignment requirements to struct drgn_architecture_info
In preparation for the next change implementing alignof(), we need to
know how base types are aligned. Add this information to struct
drgn_architecture_info. I populated this using a special C file that
doesn't compile but produces a compiler diagnostic which includes the
alignment requirements.

There may be exotic types (e.g., _BitInt, __m64, etc.) that these don't
apply to, but we'll deal with that another time.

Signed-off-by: Omar Sandoval <osandov@osandov.com>