Add alignof()
I've wanted this several times for reconstructing type information. We
need to check the DWARF information for a non-default alignment first
(continuing the layer violation from a couple of commits ago for now)
and then fall back to the default alignment rules. This is rarely
needed, so we don't precompute it or cache it.

Signed-off-by: Omar Sandoval <osandov@osandov.com>