drgndoc: allow formatting __init__() signature together with class
In all of our current code, we want to document a class separately from
its "constructor"; see commit c801e5e9b1b1 ("drgndoc: format __init__()
signature separately from class"). But for some trivial classes, like
one that will be added in an upcoming change, it makes sense to collapse
them. So, for classes with no docstring and one __init__() signature
with a docstring, document the class with the __init__() signature and
docstring.

Signed-off-by: Omar Sandoval <osandov@osandov.com>