libdrgn: dwarf_info: resolve address of inline functions with out-of-line instances
If a function is inlined at some call sites but is also compiled to a
normal function, then there will be two DW_TAG_subprogram DIEs for it:
the abstract instance root, with the DW_AT_name but no DW_AT_low_pc, and
the concrete out-of-line instance, with the DW_AT_low_pc and a
DW_AT_abstract_origin referring to the abstract instance root but no
DW_AT_name. The DWARF index finds the former but not the latter, so the
DWARF object finder returns these functions as absent objects.

To fix this, stuff DW_AT_abstract_origin attributes into the
specifications map just like we do DW_AT_declaration attributes, and
then use that to look for a concrete out-of-line instance when we index
an abstract instance root.

There was no measurable performance difference with this change.

Fixes #147.

Signed-off-by: Omar Sandoval <osandov@osandov.com>