vmtest.githubapi: add environment variable for trusting cache
Specifying a vmtest kernel version that isn't an exact match (e.g., with
a wildcard) currently always makes a GitHub API request to get the
vmtest release list. This is annoying if you have no or slow network
connectivity and don't want to check what the exact version is
(especially if you just downloaded it). Add a hack to bypass this
request by setting VMTEST_TRUST_CACHE=1 in the environment.

Signed-off-by: Omar Sandoval <osandov@osandov.com>