vmtest.config: allow adding patch version to vmtest kernel version to avoid full rebuilds
Updating the vmtest kernel version rebuilds every
architecture/flavor/version combination of the vmtest kernel. Many
changes that we make to the vmtest kernel configuration only affect some
architecture/flavor/version combinations, though. Rearrange
kconfig_localversion() so that we can add a patch level to the version
when we only need to rebuild the needed combinations.

Signed-off-by: Omar Sandoval <osandov@osandov.com>