tests: add backports of addClassCleanup() and enter{,Class}Context()
We have some test cleanup code that is annoying to do with
setUp{,Class}() and tearDown{,Class}(), and I'm about to add more.
Python 3.8 added addClassCleanup() and Python 3.11 added enterContext()
and enterClassContext() that drastically simplify test cleanup. We can
implement those for older Python versions. The only gotcha is that we
need a decorator to make sure that cleanups are run if setUpClass()
fails.

Signed-off-by: Omar Sandoval <osandov@osandov.com>