contrib: ptdrgn: remove copied run_interactive()
The new drgn.internal.repl module exports a convenient interact()
method, which accepts a banner and a namespace. So we can now drop-in
the prompt_toolkit embed function as an implementation of interact(),
rather than copying the run_interactive function and modifying it.

This does mean that the readline configuration steps from
run_interactive(), which had been stripped from this version, will be
run. But honestly, there's not much harm in configuring readline,
because prompt_toolkit won't use it anyway. It's much nicer to re-use
the existing run_interactive() function so that we don't need to modify
this script as it changes.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>