vmtest.config: skip RAID 6 algorithm benchmarking
When we load the Btrfs module for running tests, it loads the raid6_pq
module as a dependency. That module benchmarks a bunch of algorithms in
order to pick the fastest one. This adds between half a second and 2
seconds of startup time for no reason, since we don't use Btrfs with
RAID 6. Luckily, there is a kconfig option to disable benchmarking, so
enable it. Unluckily, it was added in 5.0, so also patch it in for older
vesrions.

Signed-off-by: Omar Sandoval <osandov@osandov.com>