libdrgn: program: improve checks for userspace programs
We have a bunch of places that check prog->flags to determine whether
the target is the Linux kernel, a userspace process, or a userspace core
dump. But they do so inconsistently, and the flag checks are often hard
to follow. Add a couple of helper functions,
drgn_program_is_userspace_process() and
drgn_program_is_userspace_core(), to unify these checks and make them
more clear. I left the checks in libdrgn/debug_info.c alone because
they're going to be replaced by the module API soon.

Signed-off-by: Omar Sandoval <osandov@osandov.com>