libdrgn: orc_info: refactor remove_fdes_from_orc()
In preparation for the next change, remove the special cases for keeping
ORC entries at the beginning and end. Instead, do it all in the main
loop. In practice, I only saw ~3 entries in the beginning special case,
and none at the end, so this wasn't much of an optimization anyways.

Signed-off-by: Omar Sandoval <osandov@osandov.com>