libdrgn: kdump: simplify getting the PRSTATUS attributes
Since the attribute hierarchy cannot change while iterating over the
PRSTATUS attributes, it is not necessary to take an attribute reference.

Although the attribute blob itself should not change either, it is a good
idea to keep its data pinned, because a raw pointer to it is stored in the
drgn_thread_set hash table. If some code tries to modify the PRSTATUS
attribute data, the attempt will fail with KDUMP_ERR_BUSY rather than leave
a dangling pointer in the hash table and possibly cause a UAF bug later.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>