module: add object
This allows users to get the object which the module was created from.
The primary use case is for Linux kernel modules, to return the "struct
module" associated with the drgn module object.

To simplify the implementation, we've added the restriction that
Program.linux_kernel_loadable_module(), and the associated C APIs, will
only accept objects of type "struct module *".

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>