orc_info: enable the use of built-in ORC
ORC has always been loaded from the ELF debug file. However, ORC is
present in the memory pages of kernel core dumps, so it can still be
used when the debug file is unavailable. Implement the ability to load
built-in ORC for vmlinux and kernel modules. We still prefer to load ORC
from the debug file wherever possible, because this is almost certainly
faster.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>