drgn.helpers.linux.slab: harden against cycles in freelist
Breno and Leandro debugged an issue where drgn was stuck in
identify_address() in our production crash monitoring system. They found
that _slub_get_freelist() was in an infinite loop because the freelist
contained a cycle. This indicates some sort of corruption or race.

We generally assume that data structures are valid when parsing them,
but this case is particularly risky because we use identify_address()
liberally. Luckily, we're already building a set of freelist pointers,
so we can check for a cycle relatively cheaply.

Breno and Leandro suggested breaking from the loop when a cycle is
detected, but that can cause silently undiagnosed corruptions and bogus
results from slab_cache_for_each_allocated_object() and
slab_object_info(). Instead, let's raise an exception.
slab_cache_for_each_allocated_object() has to fail hard, but
slab_object_info() can catch the exception and indicate the corruption
in its return value.

Closes #437.

Reported-by: Breno Leitao <leitao@debian.org>
Reported-by: Leandro Silva <ldfsilva@gmail.com>
Signed-off-by: Omar Sandoval <osandov@osandov.com>