orc_info: apply debug file bias to pc_base at load
When looking up CFI rules using ORC, we use module->debug_file_bias
unconditionally. This makes sense when the ORC is always loaded from an
ELF debug file. However, now that built-in ORC can be loaded, it is
possible that:

1. ORC is loaded from the built-in source, prior to loading the debug
   file. The module->debug_file_bias == 0, so the ORC is interpreted
   correctly.
2. Later, a debug file is loaded, updating debug_file_bias. However, the
   ORC hasn't been loaded from the debug file, so the bias is not
   applicable.
3. Future CFI lookups using ORC fail due to the extra bias.

To avoid this, apply the debug_file_bias once to module->orc.pc_base,
at the time we load the ORC sections out of the debug file. This ensures
that the bias is only applied to the ORC data when we know we need it.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>