orc_info: store biased pc_base
When looking up CFI rules using ORC, we apply module->debug_file_bias to
pc_base.  This made sense when the ORC was always loaded from an ELF
debug file. However, now that built-in ORC can be loaded, this only
works for ORC when debug_file_bias is zero; that is, when there is no
debug_file.

This is a problem, because it's possible that a debug_file is loaded
when built-in ORC is used. It can happen either when the debug_file has
no ORC sections present, or when the debug_file is loaded after the
built-in ORC is.

To avoid this, we define the pc_base as the biased (runtime) address of
the orc_unwind_ip section. This is already the case for built-in ORC,
but when we load it from the debug_file, we must apply the bias. In
cases such as remove_fdes_from_orc(), which want to compare ORC PCs
against unbiased addresses, they'll need to subtract the bias to match
the relevant file.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>