libdrgn: debug_info: add debuginfod client scaffolding
We currently use debuginfod via libdwfl, but when we get rid of our
libdwfl dependency, we'll need to do the debuginfod calls ourselves. So,
let's add the scaffolding for using libdebuginfod. We provide three
choices at build time:

* No debuginfod (./configure --without-debuginfod).
* Soft dependency: load libdebuginfod with dlopen at runtime if
  available (./configure --with-debuginfod --enable-dlopen-debuginfod).
  This is the default and probably what we want distros to use.
* Hard dependency: link against libdebuginfod (./configure
  --with-debuginfod --disable-dlopen-debuginfod). This is intended for
  environments where dlopen can't be used (e.g., manylinux wheels).

The client handle will be created lazily, so for now this just sets up
some wrappers and doesn't do much with them.

Signed-off-by: Omar Sandoval <osandov@osandov.com>