libdrgn: x86_64: allow phys_base to be 0
Mark reported recursive address translation errors when debugging a QEMU
guest memory dump. We're supposed to be handling this in
linux_kernel_pgtable_iterator_next_x86_64() by translating
swapper_pg_dir to a physical address. However, it decides whether
phys_base was present in the VMCOREINFO by comparing phys_base to 0, and
this dump legitimately had a phys_base of 0. Add a separate flag so we
can distinguish missing vs 0.

Reported-and-tested-by: Mark Harmstone <maharmstone@fb.com>
Fixes: 5b39bfb547c6 ("libdrgn: x86_64: avoid recursive address translation for swapper_pg_dir")
Signed-off-by: Omar Sandoval <osandov@osandov.com>