tests: linux_kernel: test built-in ORC unwinding
Loading built-in ORC is a difficult functionality to test: it is best
tested when there is no debuginfo file. Thus, we add two tests: one
simpler test in which the kernel has debuginfo, but a module does not,
and we must unwind a stack with functions from the module. The second
test is more complex, where we create a program with no debuginfo at
all, and provide it just enough data to initialize the module API and
unwind with built-in ORC.

In both cases, to verify that drgn is actually using ORC, we capture its
log messages.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>