Make StackFrame.name fall back to symbol/PC and add StackFrame.function_name
Multiple people have lamented that StackFrame.name is None for functions
implemented in assembly or missing debug info for any other reason. With
DWARFless debugging, this will be way more common. My original hope was
that StackFrame.name would strictly be the function name from the
debugging information and that callers would fall back to getting the
symbol name themselves. However, the distinction isn't super meaningful
to users, so let's add the fallback directly to StackFrame.name and add
StackFrame.function_name with the old behavior of StackFrame.name.

Signed-off-by: Omar Sandoval <osandov@osandov.com>