libdrgn: always make drgn_program available to Python
In order to support Python plugins even when using libdrgn directly, we
need every struct drgn_program to be available to Python as a
drgn.Program. So, let's rework drgn_program_create*() to always create a
drgn.Program, initializing the Python interpreter and extension module
if needed. We still support compiling libdrgn without Python support, in
which case we keep the old behavior.

Signed-off-by: Omar Sandoval <osandov@osandov.com>