libdrgn: stack_trace: give hint when stack_trace(0) fails
PID 0 is not unique in the Linux kernel; there is a task with PID 0 for
each CPU. stack_trace(0) currently fails with a generic "task not found"
error message, which can be confusing; see #462. Add a hint to use
idle_task() to the error message when the given PID is 0.

Signed-off-by: Omar Sandoval <osandov@osandov.com>