Replace debug_info_path with debug_info_options
There are more options that we will want to control other than the debug
directories, so create a more general object to represent them. libdrgn
exposes it as an opaque pointer with getters and setters (creating
minimal ABI burden), and the Python bindings expose it as a
dataclass-esque type. struct drgn_program has a default options object,
and an upcoming change will allow passing custom options to a single
finder invocation.

For now, there are three types of options: list of strings, bool, and
enum. (The single debug_info_path colon-separated string is replaced by
a list.) This uses a bunch of ugly macros to reduce boilerplate for
adding options.

This also adds a couple of command line options for controlling the
debug directories to prepare for doing the same for other options.

Signed-off-by: Omar Sandoval <osandov@osandov.com>