Add a debug info option for kernel directories
We have a very old feature request to support searching for vmlinux and
kernel modules in a directory specified by the user. This is useful, for
example, if someone is given a vmcore and the corresponding kernel
files. Now that we have the flexibility of the module and debug info
options APIs, we can finally support it and add a command line option.

Closes #17.

Signed-off-by: Omar Sandoval <osandov@osandov.com>