Support searching for kernel modules without depmod metadata
For live kernel debugging, you usually have a properly installed kernel
that has run depmod. However, for core dumps, it's common to get an
archive containing all of the kernel module files and no depmod
metadata. Peter Collingbourne also reported that the depmod situation on
Android is complicated: Android doesn't have a modules.dep.bin file, and
modules are split among different images.

Let's add support for walking kernel module directories if
modules.dep.bin isn't found and a debug info option for controlling
whether depmod, walking, or both are tried. We default to using depmod
if it is available and walking if not.

Closes #369.

Signed-off-by: Omar Sandoval <osandov@osandov.com>