libdrgn: python: pass fault errors through
Some parts of libdrgn detect drgn error codes and handle them
appropriately. For instance, the stack tracing code expects to get a
fault error. A drgn error that has been translated into a Python
exception, and back to a drgn error, no longer retains its code. This
means that if the stack tracing code is used with Python memory readers,
the fault errors will not be treated as fault errors.

For now, special case the fault errors so they are translated back to
drgn exceptions. In general, translating Python exceptions back to drgn
errors is not a great idea, because there is a loss of information. But
in this specific case, no information is lost, and it allows custom
memory readers to behave more like built-in memory readers.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>