libdrgn: stack_trace: allow unwinding anything with a pt_regs
Right now, we're a bit conservative in the stack tracing code: we
only allow unwinding userspace cores, or Linux kernel programs. These
are the only two types of programs for which we can get initial
registers. However, if the user provides a pt_regs object, then there's
no reason we can't try to do a stack trace with that.

Move the check for live programs into drgn_get_initial_registers(),
after we've already handled pt_regs.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>