Add contrib/pstack.py script for printing full kernel + user stack
This script is definitely not perfect: it assumes that the task is
running in the same filesystem namespace as this script. It also cannot
do much to handle unwinding through JIT compiled code, or anything which
doesn't have a file mapping for which we can find debuginfo (though
usually, this sort of code uses frame pointers). It tries its best to
avoid getting misled by files which have changed on-disk (e.g. if a
shared library is updated by your package manager).

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>