contrib: add pstack.py script to print kernel + user stack
This provides an example of some advanced usage of drgn to perform stack
traces of userspace processes by reading their memory through a kernel
program, or by dumping that memory and using it later to create stack
traces. Both approaches are useful.

This is all enabled by the Module API, as it lets us be very
fine-grained about exactly what DSOs are loaded, and where. For Fedora &
derivatives specifically, more symbols can be made available if
.gnu_debugdata support is available for drgn. But it's not a
requirement, and it doesn't impact the stack tracing functionality.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>