libdrgn: dwarf_info: add bias to DW_OP_addr in the middle of an expression
While evaluating a DWARF location, we currently add the debug file bias
once, at the very end. This works for DW_OP_addr operations at the end
of a location description, but if there is a DW_OP_addr operation in the
middle of an expression, it ignores the bias.

Fix it by adding the bias immediately after a DW_OP_addr (or DW_OP_addrx
or DW_OP_GNU_addr_index) operation instead.

P.S. This manifested as FaultErrors when accessing the vmap_zone_size
variable on a kernel compiled with Clang 19. The reason why is really
interesting: vmap_zone_size only has two possible values, 1 and
(1 << 4) * PAGE_SIZE. It's only used in one place:

	static inline unsigned int
	addr_to_node_id(unsigned long addr)
	{
		return (addr / vmap_zone_size) % nr_vmap_nodes;
	}

To optimize that division, Clang apparently replaced vmap_zone_size with
a shift such that the division (addr / vmap_zone_size) could be replaced
with (addr >> (shift << 4)). The location description for vmap_zone_zone
is therefore:

	DW_OP_addr 0xffffffff82350194, DW_OP_deref_size 0x1,
	DW_OP_constu 0xffff, DW_OP_mul, DW_OP_lit1, DW_OP_plus,
	DW_OP_stack_value

Fixes: 35a1af7ad690 ("libdrgn: add DWARF expression evaluation")
Signed-off-by: Omar Sandoval <osandov@osandov.com>