libdrgn: dwarf_info: fix DWARF index error handling
DWARF indexing has some awkward error handling because we can't break
out of an OpenMP parallel region and also because we need to free some
temporary data structures regardless of success or error. Each thread
has its own error, and they are consolidated at a few points. This error
handling has a couple of bugs:

1. In some loops, we check the err variable shared by all threads to
   avoid doing real work if there has been an error. This is technically
   a data race.
2. After the second pass and before merging the index maps, we set
   thread_err to err (again to short circuit the merging in all threads
   if there was an error in any thread). But we later destroy thread_err
   if err is set, which will always be true, resulting in a
   use-after-free or double-free.

We don't need to optimize for the error case at all, so let's simplify
all of this to exclusively use the per-thread error and consolidate them
only at the end.

Fixes: e0921c5bdbef ("libdrgn: don't use OpenMP tasking")
Fixes: c4a122ead686 ("libdrgn: dwarf_info: scalably index all DIEs per name")
Signed-off-by: Omar Sandoval <osandov@osandov.com>