tests: dwarfwriter: rename compile_dwarf() to create_dwarf_file()
I want to use the name compile_dwarf() for a more general method that
can return information in addition to the file data (specifically,
labels), so rename it to match create_elf_file(). While we're doing
this, rename the dies parameter to units_or_dies for clarity and make
little_endian and bits keyword-only, none of which make any difference
for current callers.

Signed-off-by: Omar Sandoval <osandov@osandov.com>