cli: extract out default_globals()
Drgn has a set of globals provided in interactive mode, but they can be
useful to other modes (e.g. for executing statements directly). They may
also be useful to client code. Make this a separate, public helper. It
does make the globals_func argument to run_interactive() a bit much, but
that's ok.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>