cli: add -e option to exec() code directly
The Python interpreter has "-c" which allows directly running code
provided as a command line argument. This is useful for quick tests,
without needing to write a script and execute it, and without needing to
run the interactive interpreter. It can also be used in scripts or
one-liners.

Unfortunately, "-c" is already used in the drgn CLI. However, this
functionality would be quite useful. Let's add it to the CLI using the
option "-e", which is short for "execute" or "exec", the underlying
Python function.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>