tests: dwarfwriter: add compile_dwarf() function that returns labels
For testing .gnu_debugaltlink files, we need a way to reference DIEs
from the .gnu_debugaltlink file in the main debug file. Add a function,
compile_dwarf(), that returns the file data as well as a dictionary of
labels (and can be extended to return more information in the future if
needed).

Signed-off-by: Omar Sandoval <osandov@osandov.com>