libdrgn: dwarf_info: read all abbreviation tables before indexing CUs
During DWARF indexing, we currently read CU abbreviation tables together
with the first indexing pass. This means that while we're indexing one
CU, some other CU's abbreviation tables may not have been read yet.
However, when we index imported units, we need the imported unit's
abbreviation table to already be read. The easiest way to ensure this is
to read all of them in a pass before the first indexing pass. This
doesn't cause a measurable performance change.

Signed-off-by: Omar Sandoval <osandov@osandov.com>