libdrgn: dwarf_info: create separate CU lookup table
Currently, we use the drgn_dwarf_info::index_cus array to look up CUs by
DIE address. This requires re-sorting the array each time the global
namespace index is updated. We do this sort as the final step of
indexing.

In order to index imported units (e.g., for gnu_debugaltlink), we need
to look up CUs during indexing, which means that we need the array
sorted sooner. But, we also want to loop over only the new CUs during
indexing, which doesn't work if sorting mixes them among the old CUs.

Resolve this by adding a separate array, index_cu_lookup, that is sorted
by address so that we don't have to reorder index_cus. The extra memory
usage is negligible (for the Linux kernel on my laptop, an extra ~80K on
top of ~180MB of anonymous memory). We also avoid a startup time
regression because it's slightly faster to sort this smaller array, and
we're able to update it in parallel with reading CUs. It might also be
slightly faster to search this smaller array, although I didn't try
measuring that.

This also brings back realloc_array() that we got rid of back in commit
a95e42ef2e47 ("libdrgn/python: use vector for
Program_load_debug_info()").

Signed-off-by: Omar Sandoval <osandov@osandov.com>