libdrgn: dwarf_info: bounds check specifications while indexing
We currently skip bounds checking DW_AT_specification and
DW_AT_abstract_origin references because it doesn't matter whether
they're in bounds: we only use them as a hash table key. However, we're
going to reuse INSN_SPECIFICATION_* for DW_AT_import in
DW_TAG_imported_unit, and that will actually need to use the reference.

Add bounds checks and replace the uintptr_t specification with a
const char * now that it's guaranteed to be a valid pointer. This
doesn't cause a measurable performance change.

Signed-off-by: Omar Sandoval <osandov@osandov.com>