elf_symtab: support multiple tables per module
The current support for ELF symbol tables assumes that there can only be
one table per module. However, symbol tables from .gnu_debugdata are
intended to supplement the dynamic symbol table found in the main
executable. Support for .gnu_debugdata will require that a module may
contain multiple tables that must each be searched.

To enable this, refactor the elf_symtab code so that it operates on
individual elf symbol tables, and can be easily extended with a new
symbol table.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>