libdrgn: python: pass fault errors through
Some parts of libdrgn detect drgn error codes and handle them
appropriately. For instance, the stack tracing code expects to get a
fault error. A drgn error that has been translated into a Python
exception, and back to a drgn error, no longer retains its code. This
means that if the stack tracing code is used with Python memory readers,
the fault errors will not be treated as fault errors.

In general, it may not be a good idea to translate every Python
exception back to a drgn error. There may be a small performance cost to
doing so, and what's more: it can be quite useful to know that a Python
error was wrapped into a drgn error. So for now, we'll make this a
special case for FaultErrors, so that custom memory readers will behave
as expected.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>