libdrgn: dwarf_info: resolve address of inline member functions
GCC and Clang both represent inline member functions that have an
out-of-line instance with a declaration in the class scope, an abstract
instance root at the top level that refers to the declaration with
DW_AT_specification, and a concrete instance at the top level that
refers to the abstract instance root with DW_AT_abstract_origin. To
index this correctly, we therefore need to resolve two references.

DWARF indexing can resolve both references: one if a DIE has
DW_AT_declaration and another if it has DW_AT_inline. The problem for
inline member functions is that the check for DW_AT_inline checks the
original (declaration) DIE, which doesn't have DW_AT_inline. So, it only
resolves the first reference but not the second.

We could store an extra flag in the specification map to indicate
whether a definition has DW_AT_inline and check that. However, that
would require additional memory.

Instead, this replaces the check for DW_AT_inline with a more general
check for the absence of DW_AT_low_pc/DW_AT_ranges. This triggers the
necessary second lookup for declaration DIEs while still working for the
global inline function case. It will also work for functions in
supplementary debug files in the future.

The second lookup may be unnecessary in some cases, but this had no
measurable performance difference.

Fixes: a8e2e1719760 ("libdrgn: dwarf_info: resolve address of inline functions with out-of-line instances")
Signed-off-by: Omar Sandoval <osandov@osandov.com>